requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
joblib
apache-airflow